# backend/api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/carts/{cartId}/share":{"post":{"tags":["shop-cart-controller"],"summary":"Share a cart with another shopper","description":"Grants sharing permission to another shopper by email","operationId":"shareCart","parameters":[{"name":"cartId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShareCartRequestDTO"}}},"required":true},"responses":{"200":{"description":"Cart shared successfully","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Failed to share cart due to invalid input or permission","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["auth-controller"],"operationId":"refreshToken","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/auth/logout-all":{"post":{"tags":["auth-controller"],"operationId":"logoutAllDevices","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/auth/google":{"post":{"tags":["auth-controller"],"operationId":"loginWithGoogle","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}},{"name":"X-Device-Id","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/carts/mine":{"get":{"tags":["shop-cart-controller"],"summary":"Get the current authenticated shopper's carts","operationId":"getMyCarts","responses":{"200":{"description":"List of carts retrieved successfully","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ShopCartDetailDTO"}}}}}}}},"/api/v1/admin/shoppers":{"get":{"tags":["admin-controller"],"operationId":"getAllShoppers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Shopper"}}}}}}}},"/api/v1/admin/shoppers/{id}":{"get":{"tags":["admin-controller"],"operationId":"getShopperById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["admin-controller"],"operationId":"deleteShopper","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/admin/shopcarts":{"get":{"tags":["admin-controller"],"operationId":"getAllShopCarts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ShopCart"}}}}}}}},"/api/v1/admin/shopcarts/{id}":{"get":{"tags":["admin-controller"],"operationId":"getShopCartById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/carts/{cartId}/share/{targetShopperId}":{"delete":{"tags":["shop-cart-controller"],"summary":"Remove sharing permission from a cart for a target shopper","operationId":"removeSharing","parameters":[{"name":"cartId","in":"path","required":true,"schema":{"type":"string"}},{"name":"targetShopperId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Sharing removed successfully","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Failed to remove sharing","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"ShareCartRequestDTO":{"type":"object","properties":{"cartId":{"type":"string","minLength":1},"targetShopperEmail":{"type":"string","minLength":1},"permission":{"type":"string","enum":["VIEW","EDIT","ADMIN"]}},"required":["cartId","permission","targetShopperEmail"]},"GroceryItem":{"type":"object","properties":{"designation":{"type":"string"},"quantity":{"type":"string"},"purchased":{"type":"boolean"}}},"SharePermissionEntry":{"type":"object","properties":{"shopperId":{"type":"string"},"permission":{"type":"string","enum":["VIEW","EDIT","ADMIN"]}}},"ShopCartDetailDTO":{"type":"object","properties":{"identifier":{"type":"string"},"name":{"type":"string"},"dateKey":{"type":"string"},"items":{"type":"array","items":{"$ref":"#/components/schemas/GroceryItem"}},"shopperIds":{"type":"array","items":{"type":"string"}},"shoppers":{"type":"array","items":{"$ref":"#/components/schemas/ShopperSummaryDTO"}},"createdBy":{"type":"string"},"sharePermissions":{"type":"array","items":{"$ref":"#/components/schemas/SharePermissionEntry"}},"state":{"type":"string","enum":["ACTIVE","SHOPPING","COMPLETED","ARCHIVED","TEMPLATE"]},"createdAt":{"type":"string","format":"date-time"},"lastModified":{"type":"string","format":"date-time"},"lastInteraction":{"type":"string","format":"date-time"},"templateName":{"type":"string"},"template":{"type":"boolean"}}},"ShopperSummaryDTO":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"name":{"type":"string"}}},"Shopper":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"name":{"type":"string"},"password":{"type":"string"},"provider":{"type":"string"},"shopCartIds":{"type":"array","items":{"type":"string"}}}},"ShopCart":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"items":{"type":"array","items":{"$ref":"#/components/schemas/GroceryItem"}},"shopperIds":{"type":"array","items":{"type":"string"}},"sharePermissions":{"type":"array","items":{"$ref":"#/components/schemas/SharePermissionEntry"}},"dateKey":{"type":"string"},"createdBy":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"lastModified":{"type":"string","format":"date-time"},"state":{"type":"string","enum":["ACTIVE","SHOPPING","COMPLETED","ARCHIVED","TEMPLATE"]},"lastInteraction":{"type":"string","format":"date-time"},"completedAt":{"type":"string","format":"date-time"},"currentShopper":{"type":"string"},"templateName":{"type":"string"},"public":{"type":"boolean"},"template":{"type":"boolean"}}}}}}